backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.1
scikit-image==0.21.0
pydantic==2.4.2
